Move breakpoint validation into pydevd
