+++ src/ptvsd/_util.py
@@ -7,7 +7,6 @@ from __future__ import print_function
 import contextlib
 import threading
 import time
-import types
 import sys
 
 import ptvsd.log
@@ -342,34 +341,3 @@ class Startable(object):
 
     def _stop(self):
         raise NotImplementedError
-
-
-try:
-    import dis
-except ImportError:
-    def get_code_lines(code):
-        raise NotImplementedError
-else:
-    def get_code_lines(code):
-        if not isinstance(code, types.CodeType):
-            path = code
-            with open(path) as f:
-                src = f.read()
-            code = compile(src, path, 'exec', 0, dont_inherit=True)
-            return get_code_lines(code)
-
-        def iterate():
-            # First, get all line starts for this code object. This does not include
-            # bodies of nested class and function definitions, as they have their
-            # own objects.
-            for _, lineno in dis.findlinestarts(code):
-                yield lineno
-
-            # For nested class and function definitions, their respective code objects
-            # are constants referenced by this object.
-            for const in code.co_consts:
-                if isinstance(const, types.CodeType) and const.co_filename == code.co_filename:
-                    for lineno in get_code_lines(const):
-                        yield lineno
-
-        return iterate()
