+++ src/ptvsd/wrapper.py
@@ -4,7 +4,6 @@
 
 from __future__ import print_function, absolute_import, unicode_literals
 
-import bisect
 import copy
 import errno
 import json
@@ -1589,27 +1588,6 @@ class VSCodeMessageProcessor(VSCLifecycleMsgProcessor):
 
         pydevd_request = copy.deepcopy(request)
         del pydevd_request['seq']  # A new seq should be created for pydevd.
-
-        # Validate breakpoints and adjust their positions.
-        args = pydevd_request['arguments']
-        if 'breakpoints' in args and 'path' in args['source']:
-            path = args['source']['path']
-            if sys.version_info < (3,) and not isinstance(path, bytes):
-                path = path.encode(sys.getfilesystemencoding())
-            path = path_to_unicode(pydevd_file_utils.norm_file_to_server(path))
-
-        try:
-            lines = sorted(_util.get_code_lines(path))
-        except Exception:
-            pass
-        else:
-            for bp in args['breakpoints']:
-                line = bp['line']
-                if line not in lines:
-                    # Adjust to the first preceding valid line.
-                    idx = bisect.bisect_left(lines, line)
-                    if idx > 0:
-                        bp['line'] = lines[idx - 1]
 
         _, _, resp_args = yield self.pydevd_request(
             pydevd_comm.CMD_SET_BREAK,
